+++ src/likelihoods/poisson.jl
@@ -10,13 +10,13 @@ for ``k ∈ N₀``, where ``θ = \\exp(f)`` and ``f`` is the latent Gaussian pro
 struct PoisLik <: Likelihood end
 
 #log of probability density
-function log_dens(poisson::PoisLik, f::VecF64, y::Vector{Int64})
+function log_dens(poisson::PoisLik, f::VecF64, y::Vector{Int})
     #where we exponentiate for positivity f = exp(fi)
     return y.*f - exp.(f) - lgamma.(1.0 .+ y)
 end
 
 #derivative of pdf wrt latent function
-function dlog_dens_df(poisson::PoisLik, f::VecF64, y::Vector{Int64})
+function dlog_dens_df(poisson::PoisLik, f::VecF64, y::Vector{Int})
     return y - exp.(f)
 end
